+++ CompetitionManagementSystem/src/main/java/com/qiaose/competitionmanagementsystem/service/CompetitionBonusService.java
@@ -16,5 +16,5 @@ public interface CompetitionBonusService{
 
     int updateByPrimaryKey(CompetitionBonus record);
 
-    CompetitionBonus selectByUserIdAndInfoId(Integer infoId, String userId);
+    CompetitionBonus selectByUserIdAndInfoId(Long infoId, String userId);
 }
